Added safe_external_pointer which exposes all features of R's external pointers
It also differs from cpp11's external_pointer in that the class represents a
reference to an *immutable* SEXP which represents the external pointer. In
contrast to cpp11's external_pointer, safe_external_pointer instances can thus
be freely and cheaply copied. While safe_external_pointer instances can be
reset(), such operations do not affect the underlying external pointer, but
instead create a new external pointer which the instance then references.

In addition to the slightly different design, safe_external_pointer is more
type-safe. When safe_external_pointer instances are created from raw SEXPs,
we check whether the SEXP points to an object of the correct type. This is
achieved by using the "tag" feature of R external pointers.

Finally, safe_external_pointer allows a "protected data" SEXP to be attached
to the external pointer. Such protected data is prevented from being GCed
for as long as the external pointer is alive. This is a feature of R's
external pointers, but not made accessible by cpp11's external_pointer.

Using safe_external_pointer in place of external_pointer fixed the crashes
that had occured.